+++ project/src/components/Map/constants.ts
@@ -1,5 +1,3 @@
-import { PointExpression } from 'leaflet';
-
 export const URL_MARKER_DEFAULT = 'img/pin.svg';
 
 export const URL_MARKER_CURRENT = 'img/pin-active.svg';
@@ -9,7 +7,3 @@ export const DEFAULT_COORDINATE = {
   longitude: 2.351499,
   zoom: 11
 };
-
-export const ICON_SIZE = [27, 39] as PointExpression | undefined;
-
-export const ICON_ANCHOR = [13.5, 39] as PointExpression | undefined;
