--- project/src/components/gallery/constants.ts
@@ -1,1 +1,0 @@
-export const MAX_PHOTOS_AMOUNT = 6;
