--- project/src/components/gallery/index.ts
@@ -1,3 +1,0 @@
-import Gallery from './gallery';
-
-export default Gallery;
