+++ project/src/components/main-empty/main-empty.tsx
@@ -1,8 +1,10 @@
+import React from 'react';
+
 type MainEmptyProps = {
   city: string;
 };
 
-const MainEmpty = ({ city }: MainEmptyProps) => (
+const MainEmpty: React.FC<MainEmptyProps> = ({ city }) => (
   <div className="cities">
     <div className="cities__places-container cities__places-container--empty container">
       <section className="cities__no-places">
