--- project/src/components/nearest-offers/index.ts
@@ -1,3 +1,0 @@
-import NearestOffers from './nearest-offers';
-
-export default NearestOffers;
