--- project/src/components/offer-component/index.ts
@@ -1,3 +1,0 @@
-import OfferComponent from './offer-component';
-
-export default OfferComponent;
