--- project/src/components/room-info/index.ts
@@ -1,3 +1,0 @@
-import RoomInfo from './room-info';
-
-export default RoomInfo;
