+++ project/src/components/sort-places/sort-places.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { SortingTypes } from '../../constants/enum';
+import { SortingTypes } from '../../constants/constants';
 import { useAppDispatch } from '../../hooks';
 import { list } from './constants';
 import { changeSort } from '../../store/offer-process/offer-process';
