--- project/src/constants/enum.ts
@@ -1,27 +1,0 @@
-export enum AuthorizationStatus {
-  Auth = 'AUTH',
-  NoAuth = 'NO_AUTH',
-  Unknown = 'UNKNOWN',
-}
-
-export enum SortingTypes {
-  PriceLowToHigh = 'PRICE_LOW_TO_HIGH',
-  PriceHighToLow = 'PRICE_HIGH_TO_LOW',
-  TopRatedFirst = 'TOP_RATED_FIRST',
-  Popular = 'POPULAR',
-}
-
-export enum NameSpace {
-  User = 'USER',
-  Offer = 'OFFER',
-  Comment = 'COMMENT',
-  NearOffers = 'NEAR_OFFERS'
-}
-
-
-export enum Status {
-  Idle = 'idle',
-  Loading = 'loading',
-  Success = 'success',
-  Error = 'error',
-}
