+++ project/src/pages/offer/offer.test.tsx
@@ -1,9 +1,9 @@
 import { createMemoryHistory } from 'history';
 import { render, screen } from '@testing-library/react';
+import {OfferComponent} from './offer';
 import { AppRoute } from '../../router/RoutePath';
-import { AuthorizationStatus } from '../../constants/enum';
+import { AuthorizationStatus } from '../../constants/constants';
 import { makeFakeOffer } from '../../utils/mocks';
-import OfferComponent from '../../components/offer-component';
 
 describe('Component: Offer', () => {
   it('should render correctly', () => {
@@ -17,8 +17,6 @@ describe('Component: Offer', () => {
         comments={[]}
         nearOtherOffers={[]}
         offerId={1}
-        offersNearbyLoading={false}
-        reviewLoaing={false}
       />
     );
