+++ project/src/services/constants.ts
@@ -4,4 +4,11 @@ export const REQUEST_TIMEOUT = 5000;
 
 export const AUTH_TOKEN_KEY_NAME = 'six-cities-token';
 
+export enum APIRoute {
+  Offers = '/hotels',
+  Login = '/login',
+  Logout = '/logout',
+  Comments = '/comments'
+}
+
 export const TIMEOUT_SHOW_ERROR = 2000;
