--- project/src/services/enum.ts
@@ -1,6 +1,0 @@
-export enum APIRoute {
-  Offers = '/hotels',
-  Login = '/login',
-  Logout = '/logout',
-  Comments = '/comments'
-}
