+++ project/src/store/comments-process/selectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
+import { NameSpace, Status } from '../../constants/constants';
 import { State } from '../../types/state';
-import { NameSpace, Status } from '../../constants/enum';
 
 export const selectStatus = (state: State) => state[NameSpace.Comment].sendCommentStatus;
 
@@ -9,5 +9,3 @@ export const selectCommentsStatus = createSelector([selectStatus], (status) => (
   isError: status === Status.Error,
   isSuccess: status === Status.Success,
 }));
-
-export const getReviewsLoading = (state: State): boolean => state[NameSpace.Comment].reviewsLoading;
