+++ project/src/store/comments-process/types.ts
@@ -1,4 +1,4 @@
-import { Status } from '../../constants/enum';
+import { Status } from '../../constants/constants';
 import { Comment } from '../../types/comments';
 import { Review } from '../../types/review';
 
@@ -6,5 +6,4 @@ export type CommentProcess = {
   loadComments: Comment[];
   nextReview: Review | null;
   sendCommentStatus: Status;
-  reviewsLoading: boolean;
 }
