+++ project/src/store/near-offers-process/api-actions.ts
@@ -1,10 +1,9 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { AppDispatch, State } from '../../types/state';
 import { AxiosInstance } from 'axios';
-import { APIRoute } from '../../services/enum';
+import { APIRoute } from '../../services/constants';
 import { Offers } from '../../types/offers';
 import { setNearOffers } from './near-offers-process';
-import { toast } from 'react-toastify';
 
 export const fetchNearOffersAction = createAsyncThunk<void, number, {
   dispatch: AppDispatch;
@@ -13,12 +12,7 @@ export const fetchNearOffersAction = createAsyncThunk<void, number, {
 }>(
   'data/fetchNearOffers',
   async (offerId, { dispatch, extra: api }) => {
-    try {
-      const { data } = await api.get<Offers[]>(`${APIRoute.Offers}/${offerId}/nearby`);
-      dispatch(setNearOffers(data));
-    } catch (e) {
-      toast.error('Unfortunately, we can\'t show nearby offers');
-      throw e;
-    }
+    const { data } = await api.get<Offers[]>(`${APIRoute.Offers}/${offerId}/nearby`);
+    dispatch(setNearOffers(data));
   },
 );
