+++ project/src/store/near-offers-process/near-offers-process.ts
@@ -1,12 +1,10 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { NameSpace } from '../../constants/enum';
+import { NameSpace } from '../../constants/constants';
 import { Offers } from '../../types/offers';
 import { NearOffersProcess } from './types';
-import { fetchNearOffersAction } from './api-actions';
 
 export const initialState: NearOffersProcess = {
   nearOffers: [],
-  offersNearbyLoading: false,
 };
 
 export const nearOffersProcessSlice = createSlice({
@@ -17,15 +15,6 @@ export const nearOffersProcessSlice = createSlice({
       state.nearOffers = action.payload;
     },
   },
-  extraReducers(builder) {
-    builder
-      .addCase(fetchNearOffersAction.pending, (state) => {
-        state.offersNearbyLoading = true;
-      })
-      .addCase(fetchNearOffersAction.fulfilled, (state) => {
-        state.offersNearbyLoading = false;
-      });
-  }
 });
 
 export const { setNearOffers } = nearOffersProcessSlice.actions;
