--- project/src/store/near-offers-process/selectors.ts
@@ -1,4 +1,0 @@
-import { NameSpace } from '../../constants/enum';
-import { State } from '../../types/state';
-
-export const getOffersNearbyLoading = (state: State): boolean => state[NameSpace.NearOffers].offersNearbyLoading;
