+++ project/src/store/near-offers-process/types.ts
@@ -2,5 +2,4 @@ import { Offers } from '../../types/offers';
 
 export type NearOffersProcess = {
   nearOffers: Offers[];
-  offersNearbyLoading: boolean;
 }
