+++ project/src/store/offer-process/api-actions.ts
@@ -1,13 +1,12 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { AppDispatch, State } from '../../types/state';
 import { AxiosInstance } from 'axios';
-import { APIRoute } from '../../services/enum';
+import { APIRoute } from '../../services/constants';
 import { loadOfferById, loadOffers, setError, setIsOffersDataLoading } from './offer-process';
 import { Offers } from '../../types/offers';
 import { OfferId } from '../../types/review';
 import { TIMEOUT_SHOW_ERROR } from '../../services/constants';
 import { store } from '..';
-import { toast } from 'react-toastify';
 
 export const clearErrorAction = createAsyncThunk(
   'data/clearError',
@@ -26,15 +25,10 @@ export const fetchOfferAction = createAsyncThunk<void, undefined, {
 }>(
   'data/fetchOffers',
   async (_arg, { dispatch, extra: api }) => {
-    try {
-      dispatch(setIsOffersDataLoading(true));
-      const { data } = await api.get<Offers[]>(APIRoute.Offers);
-      dispatch(setIsOffersDataLoading(false));
-      dispatch(loadOffers(data));
-    } catch (e) {
-      toast.error('Unfortunately, we can\'t show offers');
-      throw e;
-    }
+    dispatch(setIsOffersDataLoading(true));
+    const { data } = await api.get<Offers[]>(APIRoute.Offers);
+    dispatch(setIsOffersDataLoading(false));
+    dispatch(loadOffers(data));
   },
 );
 
@@ -45,12 +39,7 @@ export const fetchOfferByIdAction = createAsyncThunk<void, OfferId, {
 }>(
   'data/fetchOfferById',
   async ({ id }, { dispatch, extra: api }) => {
-    try {
-      const { data } = await api.get<Offers>(`${APIRoute.Offers}/${id}`);
-      dispatch(loadOfferById(data));
-    } catch (e) {
-      toast.error('Unfortunately, we can\'t show room page');
-      throw e;
-    }
+    const { data } = await api.get<Offers>(`${APIRoute.Offers}/${id}`);
+    dispatch(loadOfferById(data));
   },
 );
