+++ project/src/store/offer-process/types.ts
@@ -1,4 +1,4 @@
-import { SortingTypes } from '../../constants/enum';
+import { SortingTypes } from '../../constants/constants';
 import { Offers } from '../../types/offers';
 
 export type OfferProcess = {
